+++ frontend/src/pages/NoCodeEditor.jsx
@@ -2,854 +2,33 @@ import React, { useEffect, useRef, useState, useCallback } from 'react';
 import * as Y from 'yjs';
 import { WebsocketProvider } from 'y-websocket';
 import { useParams } from 'react-router-dom';
+
+// 모듈화된 컴포넌트들
 import ComponentLibrary from './NoCodeEditor/ComponentLibrary';
 import CanvasArea from './NoCodeEditor/CanvasArea';
 import Inspector from './NoCodeEditor/Inspector';
 import PreviewModal from './NoCodeEditor/PreviewModal';
+import EditorHeader from './NoCodeEditor/components/EditorHeader';
+import TemplateModal from './NoCodeEditor/components/TemplateModal';
+import CanvasComponent from './NoCodeEditor/components/CanvasComponent';
+import UserCursor from './NoCodeEditor/components/UserCursor';
+
+// 유틸리티 함수들
+import { 
+  GRID_SIZE, 
+  clamp, 
+  randomNickname, 
+  randomColor, 
+  getComponentDimensions,
+  resolveCollision,
+  calculateSnapLines 
+} from './NoCodeEditor/utils/editorUtils';
+
+// 컴포넌트 정의
 import { ComponentDefinitions } from './components/definitions';
-import ButtonRenderer from './NoCodeEditor/ComponentRenderers/ButtonRenderer';
-import TextRenderer from './NoCodeEditor/ComponentRenderers/TextRenderer';
-import LinkRenderer from './NoCodeEditor/ComponentRenderers/LinkRenderer';
-import AttendRenderer from './NoCodeEditor/ComponentRenderers/AttendRenderer';
-import MapView from './NoCodeEditor/ComponentEditors/MapView';
-import DdayRenderer from './NoCodeEditor/ComponentRenderers/DdayRenderer';
-import WeddingContactRenderer from './NoCodeEditor/ComponentRenderers/WeddingContactRenderer.jsx';
-import ImageRenderer from './NoCodeEditor/ComponentRenderers/ImageRenderer';
-import GridGalleryRenderer from './NoCodeEditor/ComponentRenderers/GridGalleryRenderer';
-import SlideGalleryRenderer from './NoCodeEditor/ComponentRenderers/SlideGalleryRenderer';
-import { MapInfoRenderer } from './NoCodeEditor/ComponentRenderers';
-import CalendarRenderer from './NoCodeEditor/ComponentRenderers/CalendarRenderer';
-import BankAccountRenderer from './NoCodeEditor/ComponentRenderers/BankAccountRenderer';
-import ViewportController from './NoCodeEditor/ViewportController';
-import CommentRenderer from './NoCodeEditor/ComponentRenderers/CommentRenderer';
+
 // 협업 기능 imports
 import { useCollaboration } from '../hooks/useCollaboration';
-import { LiveCursors, CollaborativeSelections } from '../components/collaboration/LiveCursors'; 
-// 그리드 크기 상수
-const GRID_SIZE = 50;
-
-// 랜덤 닉네임/색상 생성
-function randomNickname() {
-  const animals = ['Tiger', 'Bear', 'Fox', 'Wolf', 'Cat', 'Dog', 'Lion', 'Panda', 'Rabbit', 'Eagle'];
-  return animals[Math.floor(Math.random() * animals.length)] + Math.floor(Math.random() * 100);
-}
-function randomColor() {
-  const colors = ['#3B4EFF', '#FF3B3B', '#00B894', '#FDCB6E', '#6C5CE7', '#00B8D9', '#FF7675', '#636E72'];
-  return colors[Math.floor(Math.random() * colors.length)];
-}
-
-// clamp 함수 추가
-function clamp(val, min, max) {
-  return Math.max(min, Math.min(max, val));
-}
-
-// 충돌 감지 함수
-function checkCollision(comp1, comp2) {
-  const comp1Dimensions = getComponentDimensions(comp1.type);
-  const comp2Dimensions = getComponentDimensions(comp2.type);
-  
-  const comp1Width = comp1.width || comp1Dimensions.defaultWidth;
-  const comp1Height = comp1.height || comp1Dimensions.defaultHeight;
-  const comp2Width = comp2.width || comp2Dimensions.defaultWidth;
-  const comp2Height = comp2.height || comp2Dimensions.defaultHeight;
-  
-  return !(comp1.x + comp1Width <= comp2.x || 
-           comp2.x + comp2Width <= comp1.x || 
-           comp1.y + comp1Height <= comp2.y || 
-           comp2.y + comp2Height <= comp1.y);
-}
-
-// 충돌 방지 위치 계산 함수
-function resolveCollision(draggedComp, otherComponents) {
-  const COLLISION_MARGIN = 10; // 컴포넌트 간 최소 간격
-  let resolvedX = draggedComp.x;
-  let resolvedY = draggedComp.y;
-  
-  const draggedDimensions = getComponentDimensions(draggedComp.type);
-  const draggedWidth = draggedComp.width || draggedDimensions.defaultWidth;
-  const draggedHeight = draggedComp.height || draggedDimensions.defaultHeight;
-  
-  // 각 컴포넌트와의 충돌 검사 및 해결
-  for (const other of otherComponents) {
-    if (other.id === draggedComp.id) continue;
-    
-    const tempComp = { ...draggedComp, x: resolvedX, y: resolvedY };
-    if (checkCollision(tempComp, other)) {
-      const otherDimensions = getComponentDimensions(other.type);
-      const otherWidth = other.width || otherDimensions.defaultWidth;
-      const otherHeight = other.height || otherDimensions.defaultHeight;
-      
-      // 4방향 중 가장 가까운 위치로 이동
-      const moveOptions = [
-        { x: other.x - draggedWidth - COLLISION_MARGIN, y: resolvedY }, // 왼쪽
-        { x: other.x + otherWidth + COLLISION_MARGIN, y: resolvedY },   // 오른쪽
-        { x: resolvedX, y: other.y - draggedHeight - COLLISION_MARGIN }, // 위쪽
-        { x: resolvedX, y: other.y + otherHeight + COLLISION_MARGIN }   // 아래쪽
-      ];
-      
-      // 원래 위치에서 가장 가까운 옵션 선택
-      let bestOption = moveOptions[0];
-      let minDistance = Math.sqrt(Math.pow(bestOption.x - draggedComp.x, 2) + Math.pow(bestOption.y - draggedComp.y, 2));
-      
-      for (const option of moveOptions) {
-        const distance = Math.sqrt(Math.pow(option.x - draggedComp.x, 2) + Math.pow(option.y - draggedComp.y, 2));
-        if (distance < minDistance && option.x >= 0 && option.y >= 0) {
-          minDistance = distance;
-          bestOption = option;
-        }
-      }
-      
-      resolvedX = Math.max(0, bestOption.x);
-      resolvedY = Math.max(0, bestOption.y);
-    }
-  }
-  
-  return { x: resolvedX, y: resolvedY };
-}
-
-// 컴포넌트 타입별 기본 크기와 최소 크기 정의
-function getComponentDimensions(type) {
-  const dimensions = {
-    button: { defaultWidth: 120, defaultHeight: 48, minWidth: 80, minHeight: 32 },
-    text: { defaultWidth: 200, defaultHeight: 30, minWidth: 50, minHeight: 20 },
-    image: { defaultWidth: 200, defaultHeight: 150, minWidth: 50, minHeight: 50 },
-    map: { defaultWidth: 400, defaultHeight: 300, minWidth: 200, minHeight: 150 },
-    link: { defaultWidth: 150, defaultHeight: 30, minWidth: 50, minHeight: 20 },
-    attend: { defaultWidth: 300, defaultHeight: 200, minWidth: 200, minHeight: 150 },
-    dday: { defaultWidth: 200, defaultHeight: 100, minWidth: 150, minHeight: 80 },
-    weddingContact: { defaultWidth: 300, defaultHeight: 250, minWidth: 250, minHeight: 200 },
-    gridGallery: { defaultWidth: 400, defaultHeight: 300, minWidth: 200, minHeight: 200 },
-    slideGallery: { defaultWidth: 400, defaultHeight: 300, minWidth: 200, minHeight: 200 },
-    mapInfo: { defaultWidth: 300, defaultHeight: 200, minWidth: 200, minHeight: 150 },
-    calendar: { defaultWidth: 350, defaultHeight: 400, minWidth: 250, minHeight: 300 },
-    bankAccount: { defaultWidth: 300, defaultHeight: 180, minWidth: 250, minHeight: 150 },
-    comment: { defaultWidth: 300, defaultHeight: 180, minWidth: 250, minHeight: 150 }
-  };
-  return dimensions[type] || { defaultWidth: 120, defaultHeight: 40, minWidth: 80, minHeight: 30 };
-}
-
-// 캔버스 내 드래그 가능한 컴포넌트
-function CanvasComponent({ comp, selected, onSelect, onUpdate, onDelete, setSnapLines, zoom = 100, viewport = 'desktop', components = [] }) {
-  const ref = useRef();
-
-  // 더블클릭 시 텍스트 편집
-  const [editing, setEditing] = useState(false);
-  const [editValue, setEditValue] = useState(comp.props.text);
-  const [isResizing, setIsResizing] = useState(false);
-  const [isDragging, setIsDragging] = useState(false);
-  const [resizeStart, setResizeStart] = useState({ x: 0, y: 0, width: 0, height: 0, corner: '' });
-  const [dragStart, setDragStart] = useState({ x: 0, y: 0, compX: 0, compY: 0 });
-
-  // 줌 레벨에 따른 그리드 크기 계산
-  const scale = zoom / 100;
-  // 고정된 그리드 크기 사용 (줌 레벨에 관계없이 일관된 그리드)
-  const effectiveGridSize = GRID_SIZE; // 고정된 그리드 크기
-
-  const componentDimensions = getComponentDimensions(comp.type);
-  
-  // 확장된 캔버스 크기 계산 공통 함수
-  const getExtendedCanvasSize = () => {
-    const baseWidth = viewport === 'mobile' ? 375 : 1920;
-    const baseHeight = viewport === 'mobile' ? 667 : 1080;
-    
-    // 확장 컴포넌트들의 최대 Y 위치 계산
-    const extenderComponents = components?.filter(c => c.id.startsWith('canvas-extender-')) || [];
-    let maxExtendedHeight = baseHeight;
-    
-    if (extenderComponents.length > 0) {
-      const extenderMaxY = Math.max(...extenderComponents.map(c => c.y + c.height));
-      maxExtendedHeight = Math.max(baseHeight, extenderMaxY);
-    }
-    
-    return { width: baseWidth, height: maxExtendedHeight };
-  };
-  
-  // 컴포넌트별 실제 크기 계산 (props와 comp 모두 고려)
-  const getActualSize = () => {
-    // 이미지 컴포넌트의 경우 props에서 크기를 가져옴
-    if (comp.type === 'image') {
-      return {
-        width: comp.props.width || comp.width || componentDimensions.defaultWidth,
-        height: comp.props.height || comp.height || componentDimensions.defaultHeight
-      };
-    }
-    
-    // 고정 크기 컴포넌트들 (리사이즈가 어려운 컴포넌트들)
-    if (['attend', 'dday', 'weddingContact', 'calendar', 'bankAccount', 'comment'].includes(comp.type)) {
-      // 이런 컴포넌트들은 내부 레이아웃이 복잡하므로 기본 크기를 우선 사용
-      return {
-        width: comp.width || componentDimensions.defaultWidth,
-        height: comp.height || componentDimensions.defaultHeight
-      };
-    }
-    
-    // 갤러리 컴포넌트들 (동적 크기 조정 가능)
-    if (['gridGallery', 'slideGallery'].includes(comp.type)) {
-      return {
-        width: comp.width || componentDimensions.defaultWidth,
-        height: comp.height || componentDimensions.defaultHeight
-      };
-    }
-    
-    // 기본 컴포넌트들 (button, text, link 등)
-    return {
-      width: comp.width || componentDimensions.defaultWidth,
-      height: comp.height || componentDimensions.defaultHeight
-    };
-  };
-  
-  const actualSize = getActualSize();
-  const currentWidth = actualSize.width;
-  const currentHeight = actualSize.height;
-
-  useEffect(() => {
-    if (editing && ref.current) ref.current.focus();
-  }, [editing]);
-
-  // ===== 반응형 클래스 생성 =====
-  const getResponsiveClasses = (componentType) => {
-    const baseClass = `${componentType}-component`;
-    const viewportClass = `viewport-${viewport}`;
-    return `${baseClass} ${viewportClass}`;
-  };
-
-  const renderContent = () => {
-    if (editing) {
-      return (
-        <input
-          ref={ref}
-          value={editValue}
-          onChange={e => setEditValue(e.target.value)}
-          onBlur={() => { setEditing(false); onUpdate({ ...comp, props: { ...comp.props, text: editValue } }); }}
-          onKeyDown={e => {
-            if (e.key === 'Enter') {
-              setEditing(false);
-              onUpdate({ ...comp, props: { ...comp.props, text: editValue } });
-            }
-          }}
-          style={{
-            fontSize: comp.props.fontSize,
-            width: '100%',
-            border: 'none',
-            background: 'transparent',
-            outline: 'none',
-            color: 'inherit',
-            fontFamily: 'inherit',
-            fontWeight: 'inherit'
-          }}
-        />
-      );
-    }
-
-    switch (comp.type) {
-      case 'button':
-        return <ButtonRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'text':
-        return <TextRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'link':
-        return <LinkRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'attend':
-        return <AttendRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'map':
-        return <MapView {...comp.props} />;
-      case 'dday':
-        return <DdayRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'weddingContact':
-        return <WeddingContactRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'image':
-        return <ImageRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'gridGallery':
-        return <GridGalleryRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'slideGallery':
-        return <SlideGalleryRenderer comp={comp} isEditor={true} onUpdate={onUpdate} />;
-      case 'mapInfo':
-        return <MapInfoRenderer comp={comp} isEditor={true} />;
-      case 'calendar':
-        return <CalendarRenderer comp={comp} isEditor={true} />;
-      case 'bankAccount':
-        return <BankAccountRenderer comp={comp} isEditor={true} />;
-      case 'comment':
-        return <CommentRenderer comp={comp} isEditor={true} />;
-      default:
-        return <span>{comp.props.text}</span>;
-    }
-  };
-
-  // 리사이즈 핸들러
-  const handleResizeStart = (e, corner) => {
-    e.stopPropagation();
-    e.preventDefault();
-    setIsResizing(true);
-    setResizeStart({
-      x: e.clientX,
-      y: e.clientY,
-      width: currentWidth,
-      height: currentHeight,
-      corner: corner
-    });
-  };
-
-  const handleResize = (e) => {
-    if (!isResizing) return;
-    
-    const deltaX = e.clientX - resizeStart.x;
-    const deltaY = e.clientY - resizeStart.y;
-    
-    // 줌 레벨에 맞는 그리드에 스냅된 크기 계산
-    let newWidth = resizeStart.width;
-    let newHeight = resizeStart.height;
-    
-    // 모서리별 리사이즈 로직
-    switch (resizeStart.corner) {
-      case 'se':
-        newWidth = Math.max(componentDimensions.minWidth, Math.round((resizeStart.width + deltaX) / effectiveGridSize) * effectiveGridSize);
-        newHeight = Math.max(componentDimensions.minHeight, Math.round((resizeStart.height + deltaY) / effectiveGridSize) * effectiveGridSize);
-        break;
-      case 'sw':
-        newWidth = Math.max(componentDimensions.minWidth, Math.round((resizeStart.width - deltaX) / effectiveGridSize) * effectiveGridSize);
-        newHeight = Math.max(componentDimensions.minHeight, Math.round((resizeStart.height + deltaY) / effectiveGridSize) * effectiveGridSize);
-        break;
-      case 'ne':
-        newWidth = Math.max(componentDimensions.minWidth, Math.round((resizeStart.width + deltaX) / effectiveGridSize) * effectiveGridSize);
-        newHeight = Math.max(componentDimensions.minHeight, Math.round((resizeStart.height - deltaY) / effectiveGridSize) * effectiveGridSize);
-        break;
-      case 'nw':
-        newWidth = Math.max(componentDimensions.minWidth, Math.round((resizeStart.width - deltaX) / effectiveGridSize) * effectiveGridSize);
-        newHeight = Math.max(componentDimensions.minHeight, Math.round((resizeStart.height - deltaY) / effectiveGridSize) * effectiveGridSize);
-        break;
-    }
-    
-    // 확장된 캔버스 크기 계산
-    const getExtendedCanvasSize = () => {
-      const baseWidth = viewport === 'mobile' ? 375 : 1920;
-      const baseHeight = viewport === 'mobile' ? 667 : 1080;
-      
-      // 확장 컴포넌트들의 최대 Y 위치 계산
-      const extenderComponents = components?.filter(c => c.id.startsWith('canvas-extender-')) || [];
-      let maxExtendedHeight = baseHeight;
-      
-      if (extenderComponents.length > 0) {
-        const extenderMaxY = Math.max(...extenderComponents.map(c => c.y + c.height));
-        maxExtendedHeight = Math.max(baseHeight, extenderMaxY);
-      }
-      
-      return { width: baseWidth, height: maxExtendedHeight };
-    };
-    
-    const canvasSize = getExtendedCanvasSize();
-    
-    // 캔버스 경계 제한 (확장된 캔버스 크기 사용)
-    const maxWidth = Math.max(0, canvasSize.width - comp.x);
-    const maxHeight = Math.max(0, canvasSize.height - comp.y);
-    
-    newWidth = Math.min(newWidth, maxWidth);
-    newHeight = Math.min(newHeight, maxHeight);
-    
-    // 컴포넌트 타입에 따라 다르게 업데이트
-    if (comp.type === 'image') {
-      // 이미지 컴포넌트는 props에 크기 저장
-    onUpdate({
-      ...comp,
-        props: {
-          ...comp.props,
-          width: newWidth,
-          height: newHeight
-        },
-      width: newWidth,
-      height: newHeight
-    });
-    } else {
-      // 다른 컴포넌트들은 comp 레벨에 크기 저장
-      onUpdate({
-        ...comp,
-        width: newWidth,
-        height: newHeight
-      });
-    }
-  };
-
-  const handleResizeEnd = () => {
-    setIsResizing(false);
-    // 리사이즈가 끝나면 스냅라인 숨기기
-    if (setSnapLines) {
-      setSnapLines({ vertical: [], horizontal: [] });
-    }
-  };
-
-  // 드래그 핸들러
-  const handleDragStart = (e) => {
-    if (isResizing) return;
-    
-    e.stopPropagation();
-    setIsDragging(true);
-    setDragStart({
-      x: e.clientX,
-      y: e.clientY,
-      compX: comp.x,
-      compY: comp.y
-    });
-  };
-
-  const handleDrag = (e) => {
-    if (!isDragging) return;
-    
-    const deltaX = e.clientX - dragStart.x;
-    const deltaY = e.clientY - dragStart.y;
-    
-    // 확장된 캔버스 크기 계산
-    const getExtendedCanvasSize = () => {
-      const baseWidth = viewport === 'mobile' ? 375 : 1920;
-      const baseHeight = viewport === 'mobile' ? 667 : 1080;
-      
-      // 확장 컴포넌트들의 최대 Y 위치 계산
-      const extenderComponents = components?.filter(c => c.id.startsWith('canvas-extender-')) || [];
-      let maxExtendedHeight = baseHeight;
-      
-      if (extenderComponents.length > 0) {
-        const extenderMaxY = Math.max(...extenderComponents.map(c => c.y + c.height));
-        maxExtendedHeight = Math.max(baseHeight, extenderMaxY);
-      }
-      
-      return { width: baseWidth, height: maxExtendedHeight };
-    };
-    
-    const canvasSize = getExtendedCanvasSize();
-    
-    // 뷰포트에 따른 드래그 경계 제한 (확장된 캔버스 크기 사용)
-    const maxX = Math.max(0, canvasSize.width - (comp.width || componentDimensions.defaultWidth));
-    const maxY = Math.max(0, canvasSize.height - (comp.height || componentDimensions.defaultHeight));
-    
-    // 기본 위치 계산 (그리드 스냅 적용)
-    let newX = Math.round((dragStart.compX + deltaX) / effectiveGridSize) * effectiveGridSize;
-    let newY = Math.round((dragStart.compY + deltaY) / effectiveGridSize) * effectiveGridSize;
-    
-    // 다른 컴포넌트들과 스냅라인 계산
-    const tempComp = { ...comp, x: newX, y: newY };
-    const otherComponents = components?.filter(c => c.id !== comp.id) || [];
-    
-    // 스냅라인 계산
-    const snapResult = calculateSnapPosition(tempComp, otherComponents, effectiveGridSize, viewport);
-    if (snapResult.snapped) {
-      newX = snapResult.x;
-      newY = snapResult.y;
-      console.log('컴포넌트 스냅됨:', { x: newX, y: newY });
-    }
-    
-    // 충돌 방지 계산
-    const collisionResult = resolveCollision({ ...comp, x: newX, y: newY }, otherComponents);
-    newX = collisionResult.x;
-    newY = collisionResult.y;
-    
-    // 경계 제한 적용
-    newX = clamp(newX, 0, maxX);
-    newY = clamp(newY, 0, maxY);
-    
-    // 스냅라인 업데이트 (드래그 중에 실시간으로)
-    if (setSnapLines) {
-      const lines = calculateSnapLines({ ...comp, x: newX, y: newY }, otherComponents, zoom, viewport);
-      setSnapLines(lines);
-    }
-    
-    onUpdate({
-      ...comp,
-      x: newX,
-      y: newY
-    });
-  };
-
-  // 드래그 종료 핸들러 (snapLines 항상 초기화)
-  const handleDragEnd = () => {
-    setIsDragging(false);
-    // 드래그가 끝나면 snapLines를 항상 초기화 (숨김)
-    if (setSnapLines) {
-      setSnapLines({ vertical: [], horizontal: [] });
-    }
-  };
-
-  // 리사이즈 이벤트 리스너
-  useEffect(() => {
-    if (isResizing) {
-      const handleMouseMove = handleResize;
-      const handleMouseUp = handleResizeEnd;
-      
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-      
-      return () => {
-        window.removeEventListener('mousemove', handleMouseMove);
-        window.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [isResizing, resizeStart]);
-
-  // 드래그 이벤트 리스너
-  useEffect(() => {
-    if (isDragging) {
-      const handleMouseMove = handleDrag;
-      const handleMouseUp = handleDragEnd;
-      
-      window.addEventListener('mousemove', handleMouseMove);
-      window.addEventListener('mouseup', handleMouseUp);
-      
-      return () => {
-        window.removeEventListener('mousemove', handleMouseMove);
-        window.removeEventListener('mouseup', handleMouseUp);
-      };
-    }
-  }, [isDragging, dragStart]);
-
-  return (
-    <div
-      ref={ref}
-      className={`canvas-component ${getResponsiveClasses(comp.type)}`}
-      data-component-id={comp.id}
-      style={{
-        position: 'absolute',
-        left: comp.x, 
-        top: comp.y,
-        width: comp.width || componentDimensions.defaultWidth,
-        height: comp.height || componentDimensions.defaultHeight,
-        border: selected ? '2px solid #3B4EFF' : '1px solid transparent',
-        cursor: isDragging ? 'grabbing' : 'grab',
-        background: 'transparent',
-        zIndex: selected ? 10 : 1,
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-        userSelect: 'none',
-        boxSizing: 'border-box',
-        pointerEvents: 'auto',
-      }}
-      onMouseDown={handleDragStart}
-      onClick={(e) => {
-          e.stopPropagation(); 
-          onSelect(comp.id); 
-      }}
-    >
-      {renderContent()}
-      
-      {/* Figma 스타일 선택 핸들 */}
-      {selected && (
-        <>
-          {/* 모서리 리사이즈 핸들 - 실제 컴포넌트 크기에 맞게 배치 */}
-          <div
-            style={{
-              position: 'absolute',
-              top: -4 / scale,
-              left: -4 / scale,
-              width: 8 / scale,
-              height: 8 / scale,
-              background: '#3B4EFF',
-              border: `${2 / scale}px solid #fff`,
-              borderRadius: '50%',
-              cursor: 'nw-resize',
-              zIndex: 11
-            }}
-            onMouseDown={(e) => handleResizeStart(e, 'nw')}
-          />
-          <div
-            style={{
-              position: 'absolute',
-              top: -4 / scale,
-              left: currentWidth - 4 / scale,
-              width: 8 / scale,
-              height: 8 / scale,
-              background: '#3B4EFF',
-              border: `${2 / scale}px solid #fff`,
-              borderRadius: '50%',
-              cursor: 'ne-resize',
-              zIndex: 11
-            }}
-            onMouseDown={(e) => handleResizeStart(e, 'ne')}
-          />
-          <div
-            style={{
-              position: 'absolute',
-              top: currentHeight - 4 / scale,
-              left: -4 / scale,
-              width: 8 / scale,
-              height: 8 / scale,
-              background: '#3B4EFF',
-              border: `${2 / scale}px solid #fff`,
-              borderRadius: '50%',
-              cursor: 'sw-resize',
-              zIndex: 11
-            }}
-            onMouseDown={(e) => handleResizeStart(e, 'sw')}
-          />
-          <div
-            style={{
-              position: 'absolute',
-              top: currentHeight - 4 / scale,
-              left: currentWidth - 4 / scale,
-              width: 8 / scale,
-              height: 8 / scale,
-              background: '#3B4EFF',
-              border: `${2 / scale}px solid #fff`,
-              borderRadius: '50%',
-              cursor: 'se-resize',
-              zIndex: 11
-            }}
-            onMouseDown={(e) => handleResizeStart(e, 'se')}
-          />
-          
-          {/* 삭제 버튼 - 실제 컴포넌트 크기에 맞게 배치 */}
-          <button
-            onClick={e => { e.stopPropagation(); onDelete(comp.id); }}
-            style={{
-              position: 'absolute', 
-              top: -20 / scale, 
-              left: currentWidth + 4 / scale,
-              background: '#FF3B3B', 
-              color: '#fff', 
-              border: 'none', 
-              borderRadius: '50%',
-              width: 24 / scale, 
-              height: 24 / scale, 
-              cursor: 'pointer', 
-              fontWeight: 'bold',
-              fontSize: 14 / scale,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              boxShadow: `0 ${2 / scale}px ${8 / scale}px rgba(255, 59, 59, 0.3)`,
-              transition: 'all 0.2s',
-              zIndex: 12
-            }}
-            onMouseEnter={(e) => {
-              e.target.style.transform = 'scale(1.1)';
-              e.target.style.background = '#ff5252';
-            }}
-            onMouseLeave={(e) => {
-              e.target.style.transform = 'scale(1)';
-              e.target.style.background = '#FF3B3B';
-            }}
-            title="Delete"
-          >
-            ×
-          </button>
-        </>
-      )}
-    </div>
-  );
-}
-
-// 사용자 커서 표시
-function UserCursor({ x, y, color, nickname }) {
-  return (
-    <div
-      style={{
-        position: 'absolute',
-        left: x, top: y,
-        pointerEvents: 'none',
-        zIndex: 9999,
-        display: 'flex', flexDirection: 'column', alignItems: 'center'
-      }}
-    >
-      <div style={{
-        width: 16, height: 16, borderRadius: '50%',
-        background: color, border: '2px solid #fff', boxShadow: '0 1px 4px #0002'
-      }} />
-      <div style={{
-        marginTop: 2, fontSize: 12, color: color, fontWeight: 'bold',
-        background: '#fff', borderRadius: 4, padding: '2px 6px', boxShadow: '0 1px 4px #0001'
-      }}>{nickname}</div>
-    </div>
-  );
-}
-
-// 스냅 위치 계산 함수 (실제 스냅 기능 - 중앙선, 정렬, 그리드 스냅)
-function calculateSnapPosition(draggedComp, otherComponents, gridSize = 50, viewport = 'desktop') {
-  const SNAP_THRESHOLD = 12;
-  let snappedX = draggedComp.x;
-  let snappedY = draggedComp.y;
-  let snapped = false;
-
-  const draggedDimensions = getComponentDimensions(draggedComp.type);
-  const draggedWidth = draggedComp.width || draggedDimensions.defaultWidth;
-  const draggedHeight = draggedComp.height || draggedDimensions.defaultHeight;
-
-  // 드래그된 컴포넌트의 주요 위치들
-  const draggedLeft = draggedComp.x;
-  const draggedRight = draggedComp.x + draggedWidth;
-  const draggedTop = draggedComp.y;
-  const draggedBottom = draggedComp.y + draggedHeight;
-  const draggedCenterX = draggedComp.x + draggedWidth / 2;
-  const draggedCenterY = draggedComp.y + draggedHeight / 2;
-
-  // 캔버스 크기 (뷰포트에 따라)
-  const canvasWidth = viewport === 'mobile' ? 375 : 1920;
-  const canvasHeight = viewport === 'mobile' ? 667 : 1080;
-  const canvasCenterX = canvasWidth / 2;
-  const canvasCenterY = canvasHeight / 2;
-
-  // 1. 중앙선 스냅 (최우선)
-  if (Math.abs(draggedCenterX - canvasCenterX) < SNAP_THRESHOLD) {
-    snappedX = canvasCenterX - draggedWidth / 2;
-    snapped = true;
-  }
-  if (Math.abs(draggedCenterY - canvasCenterY) < SNAP_THRESHOLD) {
-    snappedY = canvasCenterY - draggedHeight / 2;
-    snapped = true;
-  }
-
-  // 2. 다른 컴포넌트들과의 정렬 스냅 체크
-  if (!snapped) {
-    for (const other of otherComponents) {
-      const otherDimensions = getComponentDimensions(other.type);
-      const otherWidth = other.width || otherDimensions.defaultWidth;
-      const otherHeight = other.height || otherDimensions.defaultHeight;
-
-      const otherLeft = other.x;
-      const otherRight = other.x + otherWidth;
-      const otherTop = other.y;
-      const otherBottom = other.y + otherHeight;
-      const otherCenterX = other.x + otherWidth / 2;
-      const otherCenterY = other.y + otherHeight / 2;
-
-      // X축 정렬 스냅 체크
-      if (Math.abs(draggedLeft - otherLeft) < SNAP_THRESHOLD) {
-        snappedX = otherLeft;
-        snapped = true;
-      } else if (Math.abs(draggedRight - otherRight) < SNAP_THRESHOLD) {
-        snappedX = otherRight - draggedWidth;
-        snapped = true;
-      } else if (Math.abs(draggedCenterX - otherCenterX) < SNAP_THRESHOLD) {
-        snappedX = otherCenterX - draggedWidth / 2;
-        snapped = true;
-      } else if (Math.abs(draggedLeft - otherRight) < SNAP_THRESHOLD) {
-        snappedX = otherRight;
-        snapped = true;
-      } else if (Math.abs(draggedRight - otherLeft) < SNAP_THRESHOLD) {
-        snappedX = otherLeft - draggedWidth;
-        snapped = true;
-      }
-
-      // Y축 정렬 스냅 체크
-      if (Math.abs(draggedTop - otherTop) < SNAP_THRESHOLD) {
-        snappedY = otherTop;
-        snapped = true;
-      } else if (Math.abs(draggedBottom - otherBottom) < SNAP_THRESHOLD) {
-        snappedY = otherBottom - draggedHeight;
-        snapped = true;
-      } else if (Math.abs(draggedCenterY - otherCenterY) < SNAP_THRESHOLD) {
-        snappedY = otherCenterY - draggedHeight / 2;
-        snapped = true;
-      } else if (Math.abs(draggedTop - otherBottom) < SNAP_THRESHOLD) {
-        snappedY = otherBottom;
-        snapped = true;
-      } else if (Math.abs(draggedBottom - otherTop) < SNAP_THRESHOLD) {
-        snappedY = otherTop - draggedHeight;
-        snapped = true;
-      }
-    }
-  }
-
-  // 3. 그리드 스냅 (우선순위가 낮음)
-  if (!snapped) {
-    const gridX = Math.round(draggedComp.x / gridSize) * gridSize;
-    const gridY = Math.round(draggedComp.y / gridSize) * gridSize;
-    
-    if (Math.abs(draggedComp.x - gridX) < SNAP_THRESHOLD / 2) {
-      snappedX = gridX;
-      snapped = true;
-    }
-    if (Math.abs(draggedComp.y - gridY) < SNAP_THRESHOLD / 2) {
-      snappedY = gridY;
-      snapped = true;
-    }
-  }
-
-  return {
-    x: snappedX,
-    y: snappedY,
-    snapped
-  };
-}
-
-// 스냅라인 계산 함수 (정렬, 간격, 그리드, 중앙선 스냅 모두 지원)
-function calculateSnapLines(draggedComp, allComponents, zoom = 100, viewport = 'desktop') {
-  const SNAP_THRESHOLD = 8;
-  // 고정된 그리드 크기 사용 (줌 레벨에 관계없이 일관된 그리드)
-  const effectiveGridSize = GRID_SIZE; // 고정된 그리드 크기
-  const snapLines = { vertical: [], horizontal: [] };
-  if (!draggedComp) return snapLines;
-
-  // 캔버스 크기 (뷰포트에 따라)
-  const canvasWidth = viewport === 'mobile' ? 375 : 1920;
-  const canvasHeight = viewport === 'mobile' ? 667 : 1080;
-
-  // 1. 중앙선 스냅 (Canvas Center)
-  const draggedDimensions = getComponentDimensions(draggedComp.type);
-  const draggedWidth = draggedComp.width || draggedDimensions.defaultWidth;
-  const draggedHeight = draggedComp.height || draggedDimensions.defaultHeight;
-  
-  const canvasCenterX = canvasWidth / 2;
-  const canvasCenterY = canvasHeight / 2;
-  const compCenterX = draggedComp.x + draggedWidth / 2;
-  const compCenterY = draggedComp.y + draggedHeight / 2;
-  
-  // 수직 중앙선 (캔버스 중앙)
-  if (Math.abs(compCenterX - canvasCenterX) < SNAP_THRESHOLD) {
-    snapLines.vertical.push({ x: canvasCenterX, type: 'center' });
-  }
-  
-  // 수평 중앙선 (캔버스 중앙)
-  if (Math.abs(compCenterY - canvasCenterY) < SNAP_THRESHOLD) {
-    snapLines.horizontal.push({ y: canvasCenterY, type: 'center' });
-  }
-
-  // 2. 정렬 스냅 (Alignment)
-  allComponents.forEach(other => {
-    if (other.id === draggedComp.id) return;
-    const otherDimensions = getComponentDimensions(other.type);
-    const otherX = [other.x, other.x + (other.width || otherDimensions.defaultWidth) / 2, other.x + (other.width || otherDimensions.defaultWidth)];
-    const dragX = [draggedComp.x, draggedComp.x + (draggedComp.width || draggedDimensions.defaultWidth) / 2, draggedComp.x + (draggedComp.width || draggedDimensions.defaultWidth)];
-    otherX.forEach(ox => {
-      dragX.forEach(dx => {
-        if (Math.abs(ox - dx) < SNAP_THRESHOLD) {
-          snapLines.vertical.push({ x: ox, type: 'align' });
-        }
-      });
-    });
-    const otherY = [other.y, other.y + (other.height || otherDimensions.defaultHeight) / 2, other.y + (other.height || otherDimensions.defaultHeight)];
-    const dragY = [draggedComp.y, draggedComp.y + (draggedComp.height || draggedDimensions.defaultHeight) / 2, draggedComp.y + (draggedComp.height || draggedDimensions.defaultHeight)];
-    otherY.forEach(oy => {
-      dragY.forEach(dy => {
-        if (Math.abs(oy - dy) < SNAP_THRESHOLD) {
-          snapLines.horizontal.push({ y: oy, type: 'align' });
-        }
-      });
-    });
-  });
-
-  // 3. 간격 스냅 (Spacing)
-  allComponents.forEach(a => {
-    allComponents.forEach(b => {
-      if (a.id === b.id || a.id === draggedComp.id || b.id === draggedComp.id) return;
-      const spacingX = Math.abs(a.x - b.x);
-      const spacingY = Math.abs(a.y - b.y);
-      if (Math.abs(Math.abs(draggedComp.x - a.x) - spacingX) < SNAP_THRESHOLD && spacingX > 0) {
-        snapLines.vertical.push({ x: draggedComp.x, type: 'spacing', spacing: spacingX });
-      }
-      if (Math.abs(Math.abs(draggedComp.y - a.y) - spacingY) < SNAP_THRESHOLD && spacingY > 0) {
-        snapLines.horizontal.push({ y: draggedComp.y, type: 'spacing', spacing: spacingY });
-      }
-    });
-  });
-
-  // 4. 그리드 스냅 (Grid) - 줌 레벨 고려
-  const gridX = Math.round(draggedComp.x / effectiveGridSize) * effectiveGridSize;
-  const gridY = Math.round(draggedComp.y / effectiveGridSize) * effectiveGridSize;
-  if (Math.abs(draggedComp.x - gridX) < SNAP_THRESHOLD) {
-    snapLines.vertical.push({ x: gridX, type: 'grid' });
-  }
-  if (Math.abs(draggedComp.y - gridY) < SNAP_THRESHOLD) {
-    snapLines.horizontal.push({ y: gridY, type: 'grid' });
-  }
-
-  return snapLines;
-}
 
 function NoCodeEditor() {
   const { roomId } = useParams();
@@ -857,7 +36,6 @@ function NoCodeEditor() {
   // 기본 상태
   const [components, setComponents] = useState([]);
   const [selectedId, setSelectedId] = useState(null);
-  const [inspector, setInspector] = useState({});
   const [snapLines, setSnapLines] = useState({ vertical: [], horizontal: [] });
   const [zoom, setZoom] = useState(100);
   const [isPreviewOpen, setIsPreviewOpen] = useState(false);
@@ -979,7 +157,7 @@ function NoCodeEditor() {
           props: { ...compDef.defaultProps }
         };
         
-        const collisionResult = resolveCollision(newComponent, components);
+        const collisionResult = resolveCollision(newComponent, components, getComponentDimensions);
         clampedX = collisionResult.x;
         clampedY = collisionResult.y;
         
@@ -1006,9 +184,9 @@ function NoCodeEditor() {
     // 협업 기능으로 컴포넌트 업데이트
     updateComponent(comp.id, comp);
       
-      // 스냅라인 계산
-    const lines = calculateSnapLines(comp, components, zoom);
-      setSnapLines(lines);
+    // 스냅라인 계산
+    const lines = calculateSnapLines(comp, components, zoom, viewport, getComponentDimensions);
+    setSnapLines(lines);
   };
 
   // 컴포넌트 삭제
@@ -1017,8 +195,6 @@ function NoCodeEditor() {
     removeComponent(id);
     if (selectedId === id) setSelectedId(null);
   };
-
-
 
   // Delete 키로 삭제
   useEffect(() => {
@@ -1029,7 +205,6 @@ function NoCodeEditor() {
     };
     window.addEventListener('keydown', onKeyDown);
     return () => window.removeEventListener('keydown', onKeyDown);
-    // eslint-disable-next-line
   }, [selectedId, components]);
 
   // 속성 인스펙터
@@ -1176,150 +351,17 @@ function NoCodeEditor() {
       background: '#fff', color: '#222', fontFamily: 'Inter, sans-serif', overflow: 'hidden'
     }}>
       {/* 에디터 헤더 */}
-      <div style={{
-        position: 'fixed',
-        top: 0,
-        left: isLibraryOpen ? 240 : 0, // ComponentLibrary 상태에 따라 동적 오프셋
-        right: selectedComp ? 340 : 0, // Inspector 너비만큼 오프셋
-        height: 60,
-        background: 'rgba(255, 255, 255, 0.95)',
-        borderBottom: '1px solid #e1e5e9',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        padding: '0 24px',
-        zIndex: 100,
-        backdropFilter: 'blur(10px)',
-        transition: 'right 0.3s cubic-bezier(0.4, 0, 0.2, 1)'
-      }}>
-        {/* 좌측: 로고와 컴포넌트 개수 */}
-        <div style={{ display: 'flex', alignItems: 'center', gap: 16 }}>
-          <h1 style={{
-            margin: 0,
-            fontSize: 20,
-            fontWeight: 700,
-            color: '#1d2129'
-          }}>
-            석재민짱
-          </h1>
-          <div style={{
-            padding: '4px 8px',
-            background: '#e3f2fd',
-            color: '#1976d2',
-            borderRadius: 4,
-            fontSize: 12,
-            fontWeight: 500
-          }}>
-            {components.length}개 컴포넌트
-          </div>
-        </div>
-
-        {/* 중앙: 뷰포트 컨트롤러 */}
-        <div style={{
-          display: 'flex',
-          justifyContent: 'center',
-          flex: 1,
-          maxWidth: selectedComp ? '300px' : '400px', // Inspector 열림 상태에 따라 조정
-          transition: 'max-width 0.3s cubic-bezier(0.4, 0, 0.2, 1)'
-        }}>
-          <ViewportController
-            currentViewport={viewport}
-            onViewportChange={handleViewportChange}
-          />
-        </div>
-
-        {/* 우측: 미리보기 버튼과 기타 */}
-        <div style={{ 
-          display: 'flex', 
-          alignItems: 'center', 
-          gap: 12,
-          minWidth: selectedComp ? '120px' : '200px', // Inspector 열림 상태에 따라 조정
-          justifyContent: 'flex-end'
-        }}>
-          {/* 템플릿 저장 버튼 (관리자만) */}
-          {isAdmin && (
-            <button
-              onClick={() => setIsTemplateSaveOpen(true)}
-              style={{
-                padding: selectedComp ? '6px 12px' : '8px 16px',
-                background: '#28a745',
-                color: '#fff',
-                border: 'none',
-                borderRadius: 8,
-                fontSize: selectedComp ? 12 : 14,
-                fontWeight: 500,
-                cursor: 'pointer',
-                display: 'flex',
-                alignItems: 'center',
-                gap: selectedComp ? 4 : 8,
-                transition: 'all 0.2s',
-                boxShadow: '0 2px 8px rgba(40, 167, 69, 0.2)',
-                whiteSpace: 'nowrap'
-              }}
-              onMouseEnter={(e) => {
-                e.target.style.background = '#218838';
-                e.target.style.transform = 'translateY(-1px)';
-                e.target.style.boxShadow = '0 4px 16px rgba(40, 167, 69, 0.3)';
-              }}
-              onMouseLeave={(e) => {
-                e.target.style.background = '#28a745';
-                e.target.style.transform = 'translateY(0)';
-                e.target.style.boxShadow = '0 2px 8px rgba(40, 167, 69, 0.2)';
-              }}
-            >
-              <span>💾</span>
-              {!selectedComp && <span>템플릿 저장</span>}
-            </button>
-          )}
-          
-          {/* 미리보기 버튼 */}
-          <button
-            onClick={() => setIsPreviewOpen(true)}
-            style={{
-              padding: selectedComp ? '6px 12px' : '8px 16px', // Inspector 열림시 작게
-              background: '#3B4EFF',
-              color: '#fff',
-              border: 'none',
-              borderRadius: 8,
-              fontSize: selectedComp ? 12 : 14, // Inspector 열림시 작게
-              fontWeight: 500,
-              cursor: 'pointer',
-              display: 'flex',
-              alignItems: 'center',
-              gap: selectedComp ? 4 : 8, // Inspector 열림시 작게
-              transition: 'all 0.2s',
-              boxShadow: '0 2px 8px rgba(59, 78, 255, 0.2)',
-              whiteSpace: 'nowrap'
-            }}
-            onMouseEnter={(e) => {
-              e.target.style.background = '#2c39d4';
-              e.target.style.transform = 'translateY(-1px)';
-              e.target.style.boxShadow = '0 4px 16px rgba(59, 78, 255, 0.3)';
-            }}
-            onMouseLeave={(e) => {
-              e.target.style.background = '#3B4EFF';
-              e.target.style.transform = 'translateY(0)';
-              e.target.style.boxShadow = '0 2px 8px rgba(59, 78, 255, 0.2)';
-            }}
-          >
-            <span>🔍</span>
-            {!selectedComp && <span>미리보기</span>} {/* Inspector 열림시 텍스트 숨김 */}
-          </button>
-
-          {/* Room ID 표시 (Inspector 열림시 숨김) */}
-          {!selectedComp && (
-            <div style={{
-              padding: '4px 8px',
-              background: '#f0f2f5',
-              borderRadius: 4,
-              fontSize: 12,
-              color: '#65676b'
-            }}>
-              Room: {roomId}
-            </div>
-          )}
-        </div>
-      </div>
+      <EditorHeader
+        components={components}
+        selectedComp={selectedComp}
+        isLibraryOpen={isLibraryOpen}
+        viewport={viewport}
+        onViewportChange={handleViewportChange}
+        onPreviewOpen={() => setIsPreviewOpen(true)}
+        onTemplateSaveOpen={() => setIsTemplateSaveOpen(true)}
+        roomId={roomId}
+        isAdmin={isAdmin}
+      />
 
       {/* 좌측: 컴포넌트 라이브러리 (토글 가능) */}
       <ComponentLibrary 
@@ -1342,38 +384,37 @@ function NoCodeEditor() {
         position: 'relative',
         overflow: 'hidden' // 내부 컴포넌트에서 스크롤 처리
       }}>
-      <CanvasArea
+        <CanvasArea
           containerRef={containerRef}
-        canvasRef={canvasRef}
-        components={components}
-        selectedId={selectedId}
+          canvasRef={canvasRef}
+          components={components}
+          selectedId={selectedId}
           users={{}} // 기존 users 대신 빈 객체
           nickname={userInfo.name}
-        snapLines={snapLines}
+          snapLines={snapLines}
           setSnapLines={setSnapLines}
-        onDrop={e => { handleDrop(e); }}
-        onDragOver={e => e.preventDefault()}
-        onClick={() => handleSelect(null)}
+          onDrop={e => { handleDrop(e); }}
+          onDragOver={e => e.preventDefault()}
+          onClick={() => handleSelect(null)}
           onMouseMove={() => {}} // 커서 추적은 협업 훅에서 처리
           onMouseUp={() => {}}
-        onSelect={handleSelect}
-        onUpdate={handleUpdate}
-        onDelete={handleDelete}
+          onSelect={handleSelect}
+          onUpdate={handleUpdate}
+          onDelete={handleDelete}
           onAddSection={handleAddSection} // 새 섹션 추가 핸들러
-        CanvasComponent={CanvasComponent}
-        UserCursor={UserCursor}
+          CanvasComponent={CanvasComponent}
+          UserCursor={UserCursor}
           zoom={zoom}
           onZoomChange={handleZoomChange}
           viewport={viewport}
           isInspectorOpen={!!selectedComp}
           isLibraryOpen={isLibraryOpen} // 라이브러리 상태 전달
-        updateCursorPosition={updateCursorPosition} // 협업 커서 위치 업데이트
-        // 협업 기능 props 추가
-        otherCursors={otherCursors}
-        otherSelections={otherSelections}
+          updateCursorPosition={updateCursorPosition} // 협업 커서 위치 업데이트
+          // 협업 기능 props 추가
+          otherCursors={otherCursors}
+          otherSelections={otherSelections}
+          getComponentDimensions={getComponentDimensions} // 컴포넌트 크기 함수
         />
-
-
       </div>
 
       {/* 우측: 속성 인스펙터 */}
@@ -1395,134 +436,16 @@ function NoCodeEditor() {
       />
       
       {/* 템플릿 저장 모달 */}
-      {isTemplateSaveOpen && (
-        <div style={{
-          position: 'fixed',
-          top: 0,
-          left: 0,
-          width: '100vw',
-          height: '100vh',
-          backgroundColor: 'rgba(0, 0, 0, 0.5)',
-          zIndex: 9999,
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center'
-        }}>
-          <div style={{
-            background: '#fff',
-            borderRadius: 12,
-            padding: 24,
-            width: 400,
-            maxWidth: '90vw',
-            boxShadow: '0 20px 60px rgba(0, 0, 0, 0.3)'
-          }}>
-            <h3 style={{ margin: '0 0 20px 0', fontSize: 18, fontWeight: 600 }}>
-              템플릿으로 저장
-            </h3>
-            
-            <div style={{ marginBottom: 16 }}>
-              <label style={{ display: 'block', marginBottom: 8, fontSize: 14, fontWeight: 500 }}>
-                템플릿 이름
-              </label>
-              <input
-                type="text"
-                value={templateData.name}
-                onChange={(e) => setTemplateData({...templateData, name: e.target.value})}
-                placeholder="템플릿 이름을 입력하세요"
-                style={{
-                  width: '100%',
-                  padding: '8px 12px',
-                  border: '1px solid #ddd',
-                  borderRadius: 6,
-                  fontSize: 14,
-                  outline: 'none'
-                }}
-              />
-            </div>
-            
-
-            
-            <div style={{ marginBottom: 16 }}>
-              <label style={{ display: 'block', marginBottom: 8, fontSize: 14, fontWeight: 500 }}>
-                카테고리
-              </label>
-              <select
-                value={templateData.category}
-                onChange={(e) => setTemplateData({...templateData, category: e.target.value})}
-                style={{
-                  width: '100%',
-                  padding: '8px 12px',
-                  border: '1px solid #ddd',
-                  borderRadius: 6,
-                  fontSize: 14,
-                  outline: 'none'
-                }}
-              >
-                <option value="wedding">웨딩</option>
-                <option value="events">이벤트</option>
-                <option value="portfolio">포트폴리오</option>
-              </select>
-            </div>
-            
-            <div style={{ marginBottom: 20 }}>
-              <label style={{ display: 'block', marginBottom: 8, fontSize: 14, fontWeight: 500 }}>
-                태그 (쉼표로 구분)
-              </label>
-              <input
-                type="text"
-                value={templateData.tags}
-                onChange={(e) => setTemplateData({...templateData, tags: e.target.value})}
-                placeholder="예: 결혼식, 초대장, 예쁜"
-                style={{
-                  width: '100%',
-                  padding: '8px 12px',
-                  border: '1px solid #ddd',
-                  borderRadius: 6,
-                  fontSize: 14,
-                  outline: 'none'
-                }}
-              />
-            </div>
-            
-            <div style={{ display: 'flex', gap: 12, justifyContent: 'flex-end' }}>
-              <button
-                onClick={() => {
-                  setIsTemplateSaveOpen(false);
-                  setTemplateData({ name: '', category: 'wedding', tags: '' });
-                }}
-                style={{
-                  padding: '8px 16px',
-                  background: '#f8f9fa',
-                  color: '#6c757d',
-                  border: '1px solid #dee2e6',
-                  borderRadius: 6,
-                  fontSize: 14,
-                  cursor: 'pointer'
-                }}
-              >
-                취소
-              </button>
-              <button
-                onClick={handleSaveAsTemplate}
-                disabled={!templateData.name}
-                style={{
-                  padding: '8px 16px',
-                  background: templateData.name ? '#28a745' : '#6c757d',
-                  color: '#fff',
-                  border: 'none',
-                  borderRadius: 6,
-                  fontSize: 14,
-                  cursor: templateData.name ? 'pointer' : 'not-allowed'
-                }}
-              >
-                저장
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-
-
+      <TemplateModal
+        isOpen={isTemplateSaveOpen}
+        onClose={() => {
+          setIsTemplateSaveOpen(false);
+          setTemplateData({ name: '', category: 'wedding', tags: '' });
+        }}
+        templateData={templateData}
+        setTemplateData={setTemplateData}
+        onSave={handleSaveAsTemplate}
+      />
 
       {/* 연결 상태 표시 */}
       {!isConnected && (
@@ -1540,8 +463,6 @@ function NoCodeEditor() {
           협업 서버 연결 중...
         </div>
       )}
-
-
 
       {/* 스타일 태그로 high-contrast, readable 스타일 보장 */}
       <style>{`
